Restructure `peer_discovery` code
