+++ ergo-rest/src/error.rs
@@ -10,3 +10,20 @@ pub enum NodeError {
     #[error("Invalid numerical URL segment")]
     InvalidNumericalUrlSegment,
 }
+
+#[derive(Debug, Error)]
+/// Peer discovery error
+pub enum PeerDiscoveryError {
+    /// `Url` error
+    #[error("URL error")]
+    UrlError,
+    /// mpsc sender error
+    #[error("MPSC sender error")]
+    MpscSender,
+    /// tokio::spawn `JoinError`
+    #[error("Join error")]
+    JoinError,
+    /// task spawn error
+    #[error("Task spawn error")]
+    TaskSpawn,
+}
